requirements.txt
streamlit
python-dotenv
google-auth
google-auth-oauthlib
google-cloud-aiplatform
google-adk
